requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
claude-agent-sdk==0.1.19
anthropic>=0.42.0
databricks-sdk>=0.40.0
python-multipart>=0.0.9
httpx>=0.27.0
